# Conflicts:
#	.gradle/4.10.1/fileHashes/fileHashes.bin
#	.gradle/4.10.1/fileHashes/fileHashes.lock
#	.gradle/4.10.1/taskHistory/taskHistory.bin
#	.gradle/4.10.1/taskHistory/taskHistory.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.idea/caches/gradle_models.ser
#	.idea/libraries/Gradle__com_appsflyer_af_android_sdk_4_10_2_aar.xml
#	.idea/workspace.xml
#	app/app.iml

+++ app/src/main/java/com/secureMarket/MarketFragment.java
@@ -684,8 +684,8 @@ public class MarketFragment extends Fragment implements
                     apksPath.mkdir();
                 }
 
-//                if (file.exists())
-//                    return file;
+                if (file.exists())
+                    return file;
                 try {
                     fileOutputStream = new FileOutputStream(file);
                     URL downloadUrl = new URL(url);
